exercícios web

+++ python/WEBnaNuvem/exerc01_3.py
@@ -1,7 +1,7 @@
 #Questão03 (exerc01_3)
 
 
-peso = int(input("Digite seu peso : "))
+peso = float(input("Digite seu peso : "))
 altura = float(input("Digite sua altura: "))
 IMC = (peso/altura**2)
 print(IMC)
